feat: global exception handler
`RestExceptionHandler` handles all exceptions thrown during execution and encapsulate them on `RestMessages`. These messages are then serialized as JSON to give a proper answer back to the caller.

A utility class called `Check` was created to hold some validation methods that, if not met, generate `RestExceptions` that are handle by the global handler. These exceptions must contain error keys that are translated to the language reported by the user (i18n).